Move Numba extra_ops tests to test_extra_ops
